cmake: Add CMake Package config file
Previously, the project's CMakeLists.txt referenced
nonexistent `cmake/LibPlumConfig.cmake`, which broke
the project installation process (`cmake --install`).

The patch adds the `cmake/LibPlumConfig.cmake.in`
package file template from which the CMake later on
generates a valid `LibPlumConfig.cmake` in the binary
directory.

The auto-generated `LibPlumConfig.cmake` is pretty simple
and just includes the `LibPlumTargets.cmake` so that the
users of the library don't need to perform extra steps
to reference the libraries after they do `find_package(LibPlum)`
in their projects.

Also, it makes sense to include `plum-static` target in the
package, so the patch also adds it to the export list.

Signed-off-by: Pavel Solodovnikov <pavel.al.solodovnikov@gmail.com>